requirements.txt
numpy==1.20.1
opencv-python==4.5.1.48
dv==1.0.10
PyAedatTools-believeinlain @ https://github.com/believeinlain/PyAedatTools/raw/import-only/dist/PyAedatTools-believeinlain-0.1.0.tar.gz
untangle==1.1.1
xmltodict==0.12.0
cython==0.29.23